# Remove unnecessary sqrt ops